Accept one node and creation of bucket notification during tenant add
