+++ cluster/credentials.go
@@ -19,8 +19,6 @@ package cluster
 import (
 	"errors"
 	"fmt"
-	"log"
-	"os"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -68,14 +66,11 @@ func createUserWithCredentials(ctx *Context, tenantShortName string, userdID uui
 		return err
 	}
 	// create minio user
-	log.Println("addMinioUser")
 	err = addMinioUser(sgt.StorageGroupTenant, tenantConf, userUICredentials.AccessKey, userUICredentials.SecretKey)
 	if err != nil {
-		log.Println("error on addMinioUser")
 		return err
 	}
 	// add readwrite canned policy
-	log.Println("addMinioCannedPolicyToUser")
 	err = addMinioCannedPolicyToUser(sgt.StorageGroupTenant, tenantConf, userUICredentials.AccessKey, "readwrite")
 	if err != nil {
 		return err
@@ -101,32 +96,6 @@ func createUserWithCredentials(ctx *Context, tenantShortName string, userdID uui
 		return err
 	}
 	return nil
-}
-
-func getPostgresNotificationMinioConfigKV() (config string) {
-	log.Println("getPostgresNotificationMinioConfigKV")
-	// Get the Database configuration
-	dbConfg := GetM3DbConfig()
-	// Build the database URL connection
-	dbConfigSSLMode := "disable"
-	if dbConfg.Ssl {
-		dbConfigSSLMode = "enable"
-	}
-	postgresTable := "bucketevents"
-	if os.Getenv("MINIO_POSTGRES_NOTIFICATION_TABLE") != "" {
-		postgresTable = os.Getenv("MINIO_POSTGRES_NOTIFICATION_TABLE")
-	}
-
-	config = fmt.Sprintf("notify_postgres state=on format=access connection_string=sslmode=%s table=%s host=%s port=%s username=%s password=%s database=%s",
-		dbConfigSSLMode,
-		postgresTable,
-		dbConfg.Host,
-		dbConfg.Port,
-		dbConfg.User,
-		dbConfg.Pwd,
-		dbConfg.Name)
-	log.Println(config)
-	return config
 }
 
 // storeUserUICredentialsSecret saves some UserUICredentials to a k8s secret on the tenant namespace
