+++ cluster/minio.go
@@ -17,9 +17,6 @@
 package cluster
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"log"
 	"os"
@@ -117,40 +114,25 @@ func removeMinioUser(sgt *StorageGroupTenant, tenantConf *TenantConfiguration, u
 
 // setMinioConfigPostgresNotification configures Minio for Postgres notification
 func setMinioConfigPostgresNotification(sgt *StorageGroupTenant, tenantConf *TenantConfiguration) error {
-	log.Println("setMinioConfig")
 	// get an admin with operator keys
 	adminClient, pErr := NewAdminClient(sgt.HTTPAddress(false), tenantConf.AccessKey, tenantConf.SecretKey)
 	if pErr != nil {
 		return pErr.Cause
 	}
 
-	// Call get config API
-	configBytes, err := adminClient.GetConfig()
+	err := adminClient.SetConfigKV(getPostgresNotificationMinioConfigKV())
 	if err != nil {
 		return tagErrorAsMinio(err)
 	}
-
-	var cfg map[string]map[string]interface{}
-
-	// Check if read data is in json format
-	if err = json.Unmarshal(configBytes, &cfg); err != nil {
-		return errors.New("Invalid JSON format: " + err.Error())
-	}
-
-	postgresConfig := getPostgresNotificationMinioConfig()
-	cfg["notify"]["postgresql"] = postgresConfig
-	log.Println(cfg)
-
-	r := bytes.NewReader(cfg)
-	err = adminClient.SetConfig(r)
+	// Restart minios after setting configuration
+	err = adminClient.ServiceRestart()
 	if err != nil {
 		return tagErrorAsMinio(err)
 	}
 	return nil
 }
 
-func getPostgresNotificationMinioConfig() map[string]map[string]interface{} {
-	log.Println("getPostgresNotificationMinioConfig")
+func getPostgresNotificationMinioConfigKV() (config string) {
 	// Get the Database configuration
 	dbConfg := GetM3DbConfig()
 	// Build the database URL connection
@@ -163,57 +145,20 @@ func getPostgresNotificationMinioConfig() map[string]map[string]interface{} {
 		postgresTable = os.Getenv("MINIO_POSTGRES_NOTIFICATION_TABLE")
 	}
 
-	// postgresJSONConfig := &target.PostgreSQLArgs{
-	// 	Enable:           true,
-	// 	Format:           "access",
-	// 	ConnectionString: fmt.Sprintf("sslmode=%s", dbConfigSSLMode),
-	// 	Table:            postgresTable,
-	// 	Host: &xnet.Host{
-	// 		Name: dbConfg.Host},
-	// 	Port:     dbConfg.Port,
-	// 	User:     dbConfg.User,
-	// 	Password: dbConfg.Pwd,
-	// 	Database: dbConfg.Name,
-	// }
-
-	// postgresJSONConfig := fmt.Sprintf(`
-	// 	{
-	// 	    "1": {
-	// 	        "enable": true,
-	// 	        "format": "access",
-	// 	        "connectionString": "sslmode=%s",
-	// 	        "table": "%s",
-	// 	        "host": "%s",
-	// 	        "port": "%s",
-	// 	        "user": "%s",
-	// 	        "password": "%s",
-	// 	        "database": "%s"
-	// 	    	}
-	// 	}`, dbConfigSSLMode,
-	// 	postgresTable,
-	// 	dbConfg.Host,
-	// 	dbConfg.Port,
-	// 	dbConfg.User,
-	// 	dbConfg.Pwd,
-	// 	dbConfg.Name)
-
-	var postgresJSONConfig map[string]map[string]interface{}
-	postgresJSONConfig["1"]["enable"] = true
-	postgresJSONConfig["1"]["format"] = "access"
-	postgresJSONConfig["1"]["connectionString"] = fmt.Sprintf("sslmode=%s", dbConfigSSLMode)
-	postgresJSONConfig["1"]["table"] = postgresTable
-	postgresJSONConfig["1"]["host"] = dbConfg.Host
-	postgresJSONConfig["1"]["port"] = dbConfg.Port
-	postgresJSONConfig["1"]["user"] = dbConfg.User
-	postgresJSONConfig["1"]["password"] = dbConfg.Pwd
-	postgresJSONConfig["1"]["database"] = dbConfg.Name
-	return postgresJSONConfig
+	config = fmt.Sprintf(`notify_postgres format=access connection_string=sslmode=%s table=%s host=%s port=%s username=%s password=%s database=%s`,
+		dbConfigSSLMode,
+		postgresTable,
+		dbConfg.Host,
+		dbConfg.Port,
+		dbConfg.User,
+		dbConfg.Pwd,
+		dbConfg.Name)
+	return config
 }
 
 // addMinioBucketNotification
-func addMinioBucketNotification(minioClient *minio.Client, bucketName string) error {
-	fmt.Println("addMinioBucketNotification")
-	queueArn := minio.NewArn("minio", "sqs", "", "_", "postgresql")
+func addMinioBucketNotification(minioClient *minio.Client, bucketName, region string) error {
+	queueArn := minio.NewArn("minio", "sqs", region, "_", "postgresql")
 	queueConfig := minio.NewNotificationConfig(queueArn)
 	queueConfig.AddEvents(minio.ObjectCreatedAll, minio.ObjectRemovedAll)
 
@@ -234,7 +179,6 @@ func tagErrorAsMinio(err error) error {
 
 // minioIsReady determines whether the MinIO for a tenant is ready or not
 func minioIsReady(ctx *Context) (bool, error) {
-	log.Println("minioIsReady")
 	// Get tenant specific MinIO client
 	minioClient, err := newTenantMinioClient(ctx, ctx.Tenant.ShortName)
 	if err != nil {
@@ -245,7 +189,6 @@ func minioIsReady(ctx *Context) (bool, error) {
 	// Check if it exist, we expect it to say no, or fail if MinIO is not ready
 	_, err = minioClient.BucketExists(randBucket)
 	if err != nil {
-		log.Println("error during bucket exists")
 		return false, tagErrorAsMinio(err)
 	}
