+++ go.mod
@@ -13,15 +13,15 @@ require (
 	github.com/grpc-ecosystem/go-grpc-middleware v1.0.0
 	github.com/grpc-ecosystem/grpc-gateway v1.12.1
 	github.com/lib/pq v1.0.0
-	github.com/minio/cli v1.21.0
-	github.com/minio/mc v0.0.0-20190908212443-54ee3a280031
-	github.com/minio/minio v0.0.0-20190920231956-112729386357
+	github.com/minio/cli v1.22.0
+	github.com/minio/mc v0.0.0-20191012041914-735aa139b19c
+	github.com/minio/minio v0.0.0-20191206215351-476111968a3f
 	github.com/minio/minio-go v0.0.0-20190327203652-5325257a208f
-	github.com/minio/minio-go/v6 v6.0.35
+	github.com/minio/minio-go/v6 v6.0.39
 	github.com/pelletier/go-toml v1.6.0
 	github.com/rs/cors v1.6.0
 	github.com/satori/go.uuid v1.2.0
-	golang.org/x/crypto v0.0.0-20190820162420-60c769a6c586
+	golang.org/x/crypto v0.0.0-20190923035154-9ee001bba392
 	golang.org/x/net v0.0.0-20191101175033-0deb6923b6d9
 	golang.org/x/oauth2 v0.0.0-20190402181905-9f3314589c9a
 	golang.org/x/sys v0.0.0-20191029155521-f43be2a4598c // indirect
